Fix module naming
