+++ Tests/GRPCProtobufCodeGenTests/ProtobufCodeGeneratorTests.swift
@@ -99,7 +99,6 @@ struct ProtobufCodeGeneratorTests {
 
         import GRPCCore
         import GRPCProtobuf
-        import SwiftProtobuf
 
         // MARK: - test.TestService
 
@@ -1103,35 +1102,6 @@ struct ProtobufCodeGeneratorTests {
       #expect(generated == expected)
     }
 
-    @Test("Generate with different module names")
-    @available(gRPCSwiftProtobuf 2.0, *)
-    func generateWithDifferentModuleNames() throws {
-      var config = ProtobufCodeGenerator.Config.defaults
-      let defaultNames = config.moduleNames
-
-      config.accessLevel = .public
-      config.indentation = 2
-      config.moduleNames.grpcCore = String(config.moduleNames.grpcCore.reversed())
-      config.moduleNames.grpcProtobuf = String(config.moduleNames.grpcProtobuf.reversed())
-      config.moduleNames.swiftProtobuf = String(config.moduleNames.swiftProtobuf.reversed())
-
-      let generator = ProtobufCodeGenerator(config: config)
-      let generated = try generator.generateCode(
-        fileDescriptor: Self.fileDescriptor,
-        protoFileModuleMappings: ProtoFileToModuleMappings(),
-        extraModuleImports: []
-      )
-
-      // Mustn't contain the default names.
-      #expect(!generated.contains(defaultNames.grpcCore))
-      #expect(!generated.contains(defaultNames.grpcProtobuf))
-      #expect(!generated.contains(defaultNames.swiftProtobuf))
-
-      // Must contain the configured names.
-      #expect(generated.contains(config.moduleNames.grpcCore))
-      #expect(generated.contains(config.moduleNames.grpcProtobuf))
-      #expect(generated.contains(config.moduleNames.swiftProtobuf))
-    }
   }
 
   @Suite("File-without-services (foo-messages.proto)")
